+++ tests/_test.py
@@ -579,7 +579,7 @@ def test_ignore_from_no_value_given(main: MockMainType) -> None:
     "kwargs,expected",
     [
         (
-            (dict(ignore_strings=[LEN_3[0]]), dict(ignore_strings=[LEN_4[0]])),
+            (dict(ignore_strings=[LEN_3[0]]), (flag.ignore_strings, LEN_4[0])),
             (
                 header()
                 + display((6, LEN_4[0]), (6, LEN_5[0]))
@@ -598,7 +598,7 @@ def test_ignore_from_no_value_given(main: MockMainType) -> None:
         (
             (
                 dict(ignore_files=[f"{templates.registered[1].name}.py"]),
-                dict(ignore_files=[f"{templates.registered[2].name}.py"]),
+                (flag.ignore_files, f"{templates.registered[2].name}.py"),
             ),
             (
                 header()
@@ -615,10 +615,9 @@ def test_ignore_from_no_value_given(main: MockMainType) -> None:
                         f"{templates.registered[1].name}.py": [LEN_5[0]]
                     }
                 ),
-                dict(
-                    ignore_from={
-                        f"{templates.registered[1].name}.py": [LEN_4[0]]
-                    }
+                (
+                    flag.ignore_from,
+                    f"{templates.registered[1].name}.py={LEN_4[0]}",
                 ),
             ),
             (
@@ -643,10 +642,9 @@ def test_ignore_from_no_value_given(main: MockMainType) -> None:
                         f"{templates.registered[1].name}.py": [LEN_5[0]]
                     }
                 ),
-                dict(
-                    ignore_from={
-                        f"{templates.registered[2].name}.py": [LEN_5[0]]
-                    }
+                (
+                    flag.ignore_from,
+                    f"{templates.registered[2].name}.py={LEN_5[0]}",
                 ),
             ),
             (
@@ -673,16 +671,16 @@ def test_ignore_from_no_value_given(main: MockMainType) -> None:
     ],
 )
 def test_ignore_from_no_override(
+    main: MockMainType,
     main_config: MockMainType,
-    main_kwargs: MockMainType,
     write_file: WriteFileType,
-    kwargs: t.Tuple[t.Dict[str, t.Any], t.Dict[str, t.Any]],
+    kwargs: t.Tuple[t.Dict[str, t.Any], t.Tuple[str, ...]],
     expected: t.Tuple[str, str],
 ) -> None:
     """Test default values aren't overridden through commandline.
 
+    :param main: Patch package entry point.
     :param main_config: Main for pyproject.toml usage..
-    :param main_kwargs: Main function for API.
     :param write_file: Create and write file.
     :param kwargs: Kwargs for main.
     :param expected: Expected result.
@@ -702,7 +700,7 @@ def test_ignore_from_no_override(
     write_file(Path.cwd() / f"{te1.name}.py", template)
     write_file(Path.cwd() / f"{te2.name}.py", template)
     result_1 = main_config(**kwargs[0])[0]
-    result_2 = main_kwargs(**kwargs[1])[0]
+    result_2 = main(*kwargs[1])[0]
     assert result_1 == expected[0]
     assert result_2 == expected[1]
 
@@ -710,7 +708,7 @@ def test_ignore_from_no_override(
 def test_file_args(main: MockMainType, write_file: WriteFileType) -> None:
     """Test passing individual file names to the ``path`` argument.
 
-    :param main: Patch package entry point.
+    :param main: patch package entry point.
     :param write_file: Create and write file.
     """
     for registered in templates.registered: